+++ src/soc/amd/common/fsp/fsp_report_resources.c
@@ -7,7 +7,7 @@
 #include <fsp/util.h>
 #include <types.h>
 
-void read_fsp_resources(struct device *dev, unsigned int *idx)
+void read_fsp_resources(struct device *dev, unsigned long *idx)
 {
 	const uint32_t mem_usable = (uintptr_t)cbmem_top();
 	const struct hob_header *hob_iterator;
